Add system_id to the station table
The station name isn't unique so the proper identifier of the station should be the pair system-station.
Also I think it might be a good idea to handle fleet carrier separately.
Also single body_id might be enough for tracking the stations but then we would need somehow to include the stars in body table but there's not the case now