Add dumb EDDN parser for prototype purpose
I'm commiting it as is since it seems to do the job of filling DB with
some data, i.e. very basic info of the system (basically x, y, z and
name) and some station info (more verbose but still very limited) and
finally able to add commodity prices to try to build route and calculate
purpose.

Surprisingly, EDDN data is tremendously not clean so it would be a whole
other pain to filter and sanitize that input further